requirements.txt
streamlit==1.31.0
plotly==5.18.0
pandas==2.2.0
numpy==1.26.3
openai==1.12.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
python-dotenv>=0.21.0 